backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
openai
python-dotenv
psycopg2-binary
gunicorn
pyjwt[crypto]
httpx
